requirements.txt
streamlit==1.38.0
python-docx==1.1.2
num2words==0.5.13
pandas==2.2.2
numpy==2.0.2
reportlab==4.2.2
pillow==10.4.0